requirements.txt
click==8.0.4
Flask==2.0.3
itsdangerous==2.1.1
Jinja2==3.0.3
jsonify==0.5
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
python-dotenv==0.19.2
Werkzeug==2.0.3
